Added OpenFeign

+++ fraud/src/main/kotlin/ru/grobikon/fraud/service/FraudCheckService.kt
@@ -1,5 +1,5 @@
 package ru.grobikon.fraud.service
 
 interface FraudCheckService {
-    fun isFraudulentCustomer(customerId: Int): Boolean
+    fun isFraudulentCustomer(customerId: Long): Boolean
 }
